+++ tests/test_custom_checkers/test_invalid_range_index_checker.py
@@ -1,0 +1,106 @@
+import astroid
+import pylint.testutils
+
+from python_ta.checkers.invalid_range_index_checker import InvalidRangeIndexChecker
+
+
+class TestInvalidRangeIndexChecker(pylint.testutils.CheckerTestCase):
+    CHECKER_CLASS = InvalidRangeIndexChecker
+
+    def test_valid_range_one_arg(self):
+        src = """
+        range(10)
+        """
+        range_node = astroid.extract_node(src)
+        with self.assertNoMessages():
+            self.checker.visit_call(range_node)
+
+    def test_valid_range_two_arg(self):
+        src = """
+        range(2, 8)
+        """
+        range_node = astroid.extract_node(src)
+        with self.assertNoMessages():
+            self.checker.visit_call(range_node)
+
+    def test_valid_range_three_arg(self):
+        src = """
+        range(2, 8, 2)
+        """
+        range_node = astroid.extract_node(src)
+        with self.assertNoMessages():
+            self.checker.visit_call(range_node)
+
+    def test_valid_range_three_arg_negative(self):
+        src = """
+        range(-10, -20, -2)
+        """
+        range_node = astroid.extract_node(src)
+        with self.assertNoMessages():
+            self.checker.visit_call(range_node)
+
+    def test_invalid_range_empty(self):
+        src = """
+        range()
+        """
+        range_node = astroid.extract_node(src)
+        with self.assertAddsMessages(
+            pylint.testutils.MessageTest(
+                msg_id="invalid-range-index",
+                node=range_node,
+                args="2",
+            ),
+            ignore_position=True,
+        ):
+            self.checker.visit_call(range_node)
+
+    def test_invalid_range_one_arg(self):
+        src = """
+        range(-10)
+        """
+        range_node = astroid.extract_node(src)
+        with self.assertAddsMessages(
+            pylint.testutils.MessageTest(
+                msg_id="invalid-range-index",
+                node=range_node,
+                args="2",
+            ),
+            ignore_position=True,
+        ):
+            self.checker.visit_call(range_node)
+
+    def test_invalid_range_two_arg(self):
+        src = """
+        range(5, 6)
+        """
+        range_node = astroid.extract_node(src)
+        with self.assertAddsMessages(
+            pylint.testutils.MessageTest(
+                msg_id="invalid-range-index",
+                node=range_node,
+                args="2",
+            ),
+            ignore_position=True,
+        ):
+            self.checker.visit_call(range_node)
+
+    def test_invalid_range_three_arg(self):
+        src = """
+        range(2, 15, 20)
+        """
+        range_node = astroid.extract_node(src)
+        with self.assertAddsMessages(
+            pylint.testutils.MessageTest(
+                msg_id="invalid-range-index",
+                node=range_node,
+                args="2",
+            ),
+            ignore_position=True,
+        ):
+            self.checker.visit_call(range_node)
+
+
+if __name__ == "__main__":
+    import pytest
+
+    pytest.main(["test_invalid_range_index_checker.py"])